Update Boston murders data
The Boston murders dataset consisted of a zip file containing a single CSV that was inadvertently buried in several layers of subdirectories. The subdirectories have been removed and the CSV file moved to the top (and now only) level of the zip file, along with a README.txt file explaining the change.