Harmonise tutorial instructions with book version
In the early tutorials there are periodic references to the `Run Code` button and other UI features that aren't present in the book version of the tutorials. So that these can be taken out while making minimal changes to the text, this commit makes a few minor changes to the wording of this tutorial.